+++ examples/plot_rfe_with_cross_validation.py
@@ -12,7 +12,7 @@ from sklearn.svm import SVC
 from sklearn.cross_validation import StratifiedKFold
 from sklearn.feature_selection import RFECV
 from sklearn.datasets import make_classification
-from sklearn.metrics import zero_one
+from sklearn.metrics import zero_one_loss
 
 # Build a classification task using 3 informative features
 X, y = make_classification(n_samples=1000, n_features=25, n_informative=3,
@@ -22,7 +22,7 @@ X, y = make_classification(n_samples=1000, n_features=25, n_informative=3,
 # Create the RFE object and compute a cross-validated score.
 svc = SVC(kernel="linear")
 rfecv = RFECV(estimator=svc, step=1, cv=StratifiedKFold(y, 2),
-              loss_func=zero_one)
+              loss_func=zero_one_loss)
 rfecv.fit(X, y)
 
 print "Optimal number of features : %d" % rfecv.n_features_